Fix #49.  Retain Ad Hoc Options during Order Merge
Previously ad hoc option types would be unintentionally discarded
from one of the variants during an order merge.

This would happen when a user had a cart when logged in, and a cart
non-logged in cart.  If the user then logs in, the carts are merged
in to a single logged in cart.

The merging process would ignore the ad hoc options and product
customizations added, creating a product with no ad hoc options even
if the product is required to have them.

Change-Id: I701ce1c8c8880e5a16f48ac71290d2b79ebd51a0